decentralization for Admin in "Account" function to delete accounts

+++ src/Controller/AdminController.php
@@ -324,11 +324,17 @@ class AdminController extends AbstractController
      * @Route("/accounts/delete/{id}",name="account_delete",requirements={"id"="\d+"})
      */
     
-     public function deleteAcc(Request $request, User $u): Response
+     public function deleteAcc(Request $request, Security $security, User $u): Response
      {
+        if ($security->isGranted('ROLE_ADMIN')) {
          $this->urepo->remove($u,true);
          return $this->redirectToRoute('account_show', [], Response::HTTP_SEE_OTHER);
      }
+     else{
+        return $this->render("error_admin.html.twig",[
+        ]);
+    }
+}
 
 }
